Moved TransferDirection from transfer to transfer::config
