+++ src/dma/transfer/config.rs
@@ -4,12 +4,10 @@ use super::super::stream::config::{
     M0a, M1a, MBurst, MSize, Minc, Ndt, NotM2MConf, PBurstConf, Pa, Pinc,
     PincConf, Pincos, TransferDirectionConf, TransferModeConf,
 };
-use super::buffer::{IncrementedBuffer, MemoryBufferType};
-use super::{
-    Buffer, IPayloadSize, Payload, PayloadSize, TransferBuffers,
-    TransferDirection,
-};
+use super::buffer::{IncrementedBuffer, MemoryBuffer, MemoryBufferType};
+use super::{Buffer, IPayloadSize, Payload, PayloadSize, TransferBuffers};
 use core::convert::TryFrom;
+use enum_as_inner::EnumAsInner;
 
 #[derive(Debug)]
 pub struct Config<'wo, Peripheral, Memory>
@@ -287,6 +285,242 @@ where
         }
 
         // Check Incrementing
+    }
+}
+
+#[derive(Debug, EnumAsInner)]
+pub enum TransferDirection<'wo, Peripheral, Memory>
+where
+    Peripheral: Payload,
+    Memory: Payload,
+{
+    P2M(PeripheralToMemory<'wo, Peripheral, Memory>),
+    M2P(MemoryToPeripheral<'wo, Peripheral, Memory>),
+    M2M(MemoryToMemory<'wo, Peripheral, Memory>),
+}
+
+impl<'wo, Peripheral, Memory> TransferDirection<'wo, Peripheral, Memory>
+where
+    Peripheral: Payload,
+    Memory: Payload,
+{
+    pub fn buffers(&self) -> &TransferBuffers<'wo, Peripheral, Memory> {
+        match self {
+            TransferDirection::P2M(p2m) => p2m.buffers(),
+            TransferDirection::M2P(m2p) => m2p.buffers(),
+            TransferDirection::M2M(m2m) => m2m.buffers(),
+        }
+    }
+
+    pub fn buffers_mut<F>(&mut self, op: F)
+    where
+        for<'a> F: FnOnce(&'a mut TransferBuffers<'wo, Peripheral, Memory>),
+    {
+        match self {
+            TransferDirection::P2M(p2m) => p2m.buffers_mut(op),
+            TransferDirection::M2P(m2p) => m2p.buffers_mut(op),
+            TransferDirection::M2M(m2m) => m2m.buffers_mut(op),
+        }
+    }
+
+    pub unsafe fn buffers_mut_unchecked(
+        &mut self,
+    ) -> &mut TransferBuffers<'wo, Peripheral, Memory> {
+        match self {
+            TransferDirection::P2M(p2m) => p2m.buffers_mut_unchecked(),
+            TransferDirection::M2P(m2p) => m2p.buffers_mut_unchecked(),
+            TransferDirection::M2M(m2m) => m2m.buffers_mut_unchecked(),
+        }
+    }
+
+    pub fn free(self) -> TransferBuffers<'wo, Peripheral, Memory> {
+        match self {
+            TransferDirection::P2M(p2m) => p2m.free(),
+            TransferDirection::M2P(m2p) => m2p.free(),
+            TransferDirection::M2M(m2m) => m2m.free(),
+        }
+    }
+}
+
+#[derive(Debug)]
+pub struct PeripheralToMemory<'wo, Peripheral, Memory>
+where
+    Peripheral: Payload,
+    Memory: Payload,
+{
+    buffers: TransferBuffers<'wo, Peripheral, Memory>,
+}
+
+impl<'wo, Peripheral, Memory> PeripheralToMemory<'wo, Peripheral, Memory>
+where
+    Peripheral: Payload,
+    Memory: Payload,
+{
+    pub fn new(buffers: TransferBuffers<'wo, Peripheral, Memory>) -> Self {
+        let s = Self { buffers };
+
+        s.check_self();
+
+        s
+    }
+
+    pub fn buffers(&self) -> &TransferBuffers<'wo, Peripheral, Memory> {
+        &self.buffers
+    }
+
+    pub fn buffers_mut<F>(&mut self, op: F)
+    where
+        for<'a> F: FnOnce(&'a mut TransferBuffers<'wo, Peripheral, Memory>),
+    {
+        op(&mut self.buffers);
+
+        self.check_self();
+    }
+
+    pub unsafe fn buffers_mut_unchecked(
+        &mut self,
+    ) -> &mut TransferBuffers<'wo, Peripheral, Memory> {
+        &mut self.buffers
+    }
+
+    pub fn free(self) -> TransferBuffers<'wo, Peripheral, Memory> {
+        self.buffers
+    }
+
+    fn check_self(&self) {
+        assert!(self.buffers.get().peripheral_buffer.is_read());
+        assert!(self.buffers.get().memory_buffer.is_read());
+    }
+}
+
+#[derive(Debug)]
+pub struct MemoryToPeripheral<'wo, Peripheral, Memory>
+where
+    Peripheral: Payload,
+    Memory: Payload,
+{
+    buffers: TransferBuffers<'wo, Peripheral, Memory>,
+}
+
+impl<'wo, Peripheral, Memory> MemoryToPeripheral<'wo, Peripheral, Memory>
+where
+    Peripheral: Payload,
+    Memory: Payload,
+{
+    pub fn new(buffers: TransferBuffers<'wo, Peripheral, Memory>) -> Self {
+        let s = Self { buffers };
+
+        s.check_self();
+
+        s
+    }
+
+    pub fn buffers(&self) -> &TransferBuffers<'wo, Peripheral, Memory> {
+        &self.buffers
+    }
+
+    pub fn buffers_mut<F>(&mut self, op: F)
+    where
+        for<'a> F: FnOnce(&'a mut TransferBuffers<'wo, Peripheral, Memory>),
+    {
+        op(&mut self.buffers);
+
+        self.check_self();
+    }
+
+    pub unsafe fn buffers_mut_unchecked(
+        &mut self,
+    ) -> &mut TransferBuffers<'wo, Peripheral, Memory> {
+        &mut self.buffers
+    }
+
+    pub fn free(self) -> TransferBuffers<'wo, Peripheral, Memory> {
+        self.buffers
+    }
+
+    fn check_self(&self) {
+        assert!(self.buffers.get().peripheral_buffer.is_write());
+        assert!(self.buffers.get().memory_buffer.is_read());
+    }
+}
+
+#[derive(Debug)]
+pub struct MemoryToMemory<'wo, Source, Dest>
+where
+    Source: Payload,
+    Dest: Payload,
+{
+    buffers: TransferBuffers<'wo, Source, Dest>,
+}
+
+impl<'wo, Source, Dest> MemoryToMemory<'wo, Source, Dest>
+where
+    Source: Payload,
+    Dest: Payload,
+{
+    pub fn new(buffers: TransferBuffers<'wo, Source, Dest>) -> Self {
+        let s = Self { buffers };
+
+        s.check_self();
+
+        s
+    }
+
+    pub fn buffers(&self) -> &TransferBuffers<'wo, Source, Dest> {
+        &self.buffers
+    }
+
+    pub fn buffers_mut<F>(&mut self, op: F)
+    where
+        for<'a> F: FnOnce(&'a mut TransferBuffers<'wo, Source, Dest>),
+    {
+        op(&mut self.buffers);
+
+        self.check_self();
+    }
+
+    pub unsafe fn buffers_mut_unchecked(
+        &mut self,
+    ) -> &mut TransferBuffers<'wo, Source, Dest> {
+        &mut self.buffers
+    }
+
+    pub fn source_buffer(&self) -> &Buffer<'wo, Source> {
+        &self.buffers.get().peripheral_buffer
+    }
+
+    pub fn dest_buffer(&self) -> &MemoryBuffer<Dest> {
+        &self.buffers.get().memory_buffer.as_single_buffer().unwrap()
+    }
+
+    pub fn dest_buffer_mut<F>(&mut self, op: F)
+    where
+        for<'a> F: FnOnce(&'a mut MemoryBuffer<Dest>),
+    {
+        self.buffers.get_mut(move |b| {
+            op(&mut b.memory_buffer.as_single_buffer_mut().unwrap())
+        });
+    }
+
+    pub unsafe fn dest_buffer_mut_unchecked(
+        &mut self,
+    ) -> &mut MemoryBuffer<Dest> {
+        self.buffers
+            .get_mut_unchecked()
+            .memory_buffer
+            .as_single_buffer_mut()
+            .unwrap()
+    }
+
+    pub fn free(self) -> TransferBuffers<'wo, Source, Dest> {
+        self.buffers
+    }
+
+    fn check_self(&self) {
+        assert!(self.buffers.get().peripheral_buffer.is_read());
+        assert!(self.buffers.get().memory_buffer.is_write());
+
+        assert!(self.buffers.get().memory_buffer.is_single_buffer());
     }
 }
